+++ aeronet/main/src/http-server-lifecycle.cpp
@@ -208,11 +208,13 @@ SingleHttpServer::SingleHttpServer(SingleHttpServer&& other)
 #endif
       _router(std::move(other._router)),
       _connections(std::move(other._connections)),
+      _keepAliveDeadlines(std::move(other._keepAliveDeadlines)),
       _sharedBuffers(std::move(other._sharedBuffers)),
       _telemetry(std::move(other._telemetry)),
       _internalHandle(std::move(other._internalHandle)),
       _lifecycleTracker(std::move(other._lifecycleTracker)),
-      _pendingReadFds(std::move(other._pendingReadFds))
+      _pendingReadFds(std::move(other._pendingReadFds)),
+      _connectionSweepState(std::exchange(other._connectionSweepState, {}))
 #ifdef AERONET_ENABLE_OPENSSL
       ,
       _tls(std::move(other._tls))
@@ -249,11 +251,13 @@ SingleHttpServer& SingleHttpServer::operator=(SingleHttpServer&& other) {
 #endif
     _router = std::move(other._router);
     _connections = std::move(other._connections);
+    _keepAliveDeadlines = std::move(other._keepAliveDeadlines);
     _sharedBuffers = std::move(other._sharedBuffers);
     _telemetry = std::move(other._telemetry);
     _internalHandle = std::move(other._internalHandle);
     _lifecycleTracker = std::move(other._lifecycleTracker);
     _pendingReadFds = std::move(other._pendingReadFds);
+    _connectionSweepState = std::exchange(other._connectionSweepState, {});
 #ifdef AERONET_ENABLE_OPENSSL
     _tls = std::move(other._tls);
 #endif